libft.h

+++ Projects/Libft/ft_strlcat.c
@@ -6,14 +6,13 @@
 /*   By: joao-vri <joao-vri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 14:49:29 by joao-vri          #+#    #+#             */
-/*   Updated: 2024/04/22 14:39:00 by joao-vri         ###   ########.fr       */
+/*   Updated: 2024/04/24 15:30:34 by joao-vri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// #include <stdio.h>
-// #include <string.h>
+#include "libft.h"
 
-int	ft_strlcat(char *dst, const char *src, size_t size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
 	size_t	len;
 	size_t	i;
@@ -37,7 +36,7 @@ int	ft_strlcat(char *dst, const char *src, size_t size)
 	const char *src = "World";
 	size_t size = 9;
 
-	printf("%i\n", ft_strlcat(dst, src, size));
+	printf("%lu\n", ft_strlcat(dst, src, size));
 	ft_strlcat(dst, src, size);
 	printf("%s", dst);
 } */
